Improve accessibility of popups for screen readers.
When a link is clicked, focus is moved into the popup, so screen
readers can access the content of the note. When the close button
(now an actual <button> with an ARIA label) is clicked, focus
returns to the link which triggered the popup.

These popups still aren't robustly accessible, but they're an
improvement.